requirements.txt
requests
itchat
pillow